refactor(be): [Issue #100-16] integration test をドメイン別に分割
635行の app.integration.test.ts を auth / receipt / stats / category に分割し、
helpers を __tests__/integration/helpers/ へ移動。test:integration の実行パスを更新。
