+++ backend/package.json
@@ -11,7 +11,7 @@
     "prisma:seed": "prisma db seed",
     "prisma:sync-sequences": "npx tsx prisma/run-sync-sequences.ts",
     "test": "vitest run",
-    "test:integration": "RUN_API_INTEGRATION=1 vitest run src/app.integration.test.ts"
+    "test:integration": "RUN_API_INTEGRATION=1 vitest run src/__tests__/integration"
   },
   "dependencies": {
     "@google/generative-ai": "^0.24.1",
